+++ changelog.md
@@ -5,6 +5,8 @@
 - Database : ajout de la possibilité de spécifier un ordre par défaut.
 - Foundation/Console : ajout de la commande artisan "optimize:clear".
 - Foundation/Console : suppression de "clear-compiled" de la commande "optimize:all".
+- Foundation/Console : affichage de la trace si exception catched dans la commande "migrate:test".
+- Nettoyage commentaires.
 
 ## 1.1.0 (2015-10-26)